requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib
python-multipart
python-dotenv
python-jose[cryptography]
psycopg2-binary
